+++ src/ast/printer/plugins/AttributeListPrintPlugin.ts
@@ -1,4 +1,4 @@
-import { map, pipe } from '../../../common/index.js';
+import { map, pipe } from '../../../utils/index.js';
 import { ASTNode, AttributeListASTNode } from '../../types.js';
 import { PrintPlugin } from '../types.js';
 import { endOfLine, joinBy, indent, wrapByPair } from './utils/index.js';
