+++ src/ast/printer/plugins/NodePrintPlugin.ts
@@ -1,4 +1,4 @@
-import { pipe, map } from '../../../common/index.js';
+import { pipe, map } from '../../../utils/index.js';
 import { NodeASTNode } from '../../types.js';
 import { PrintPlugin } from '../types.js';
 import { endOfLine, joinBy, indent, wrapByPair } from './utils/index.js';
