+++ src/ast/printer/plugins/NodeRefGroupPrintPlugin.ts
@@ -1,4 +1,4 @@
-import { pipe, map } from '../../../common/index.js';
+import { pipe, map } from '../../../utils/index.js';
 import { NodeRefGroupASTNode } from '../../types.js';
 import { PrintPlugin } from '../types.js';
 import { joinBy, wrapByPair } from './utils/index.js';
