+++ src/common/index.ts
@@ -1,4 +1,3 @@
-export * from './utils/index.js';
 export * from './type/index.js';
 export * from './attribute/index.js';
 export * from './models.js';
