+++ src/utils/functional.ts
